remove useless method
