--- SpiffWorkflow/util/impl.py
@@ -1,26 +1,0 @@
-# This file is part of SpiffWorkflow.
-#
-# SpiffWorkflow is free software; you can redistribute it and/or
-# modify it under the terms of the GNU Lesser General Public
-# License as published by the Free Software Foundation; either
-# version 3.0 of the License, or (at your option) any later version.
-#
-# SpiffWorkflow is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public
-# License along with this library; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-# 02110-1301  USA
-
-import sys
-
-
-def get_class(full_class_name):
-    parts = full_class_name.rsplit('.', 1)
-    module_name = parts[0]
-    class_name = parts[1]
-    __import__(module_name)
-    return getattr(sys.modules[module_name], class_name)
