+++ tests/SpiffWorkflow/camunda/MessageBoundaryEventTest.py
@@ -33,7 +33,7 @@ class MessageBoundaryTest(BaseTestCase):
         for step in steps:
             for task in ready_tasks:
                 if task.task_spec.name == step[0]:
-                    task.update_data(step[1])
+                    task.set_data(**step[1])
 
                 self.workflow.run_task_from_id(task.id)
                 self.workflow.do_engine_steps()
