+++ tests/SpiffWorkflow/camunda/ResetTokenNestedParallelTest.py
@@ -70,7 +70,7 @@ class ResetTokenTestNestedParallel(BaseTestCase):
             if firsttaskid is None and step['taskname']=='FormB2':
                 firsttaskid = task.id
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
@@ -89,7 +89,7 @@ class ResetTokenTestNestedParallel(BaseTestCase):
         for step in steps:
             task = self.get_ready_user_tasks()[0]
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
@@ -147,7 +147,7 @@ class ResetTokenTestNestedParallel(BaseTestCase):
             if firsttaskid is None and step['taskname']=='FormA2':
                 firsttaskid = task.id
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
@@ -182,7 +182,7 @@ class ResetTokenTestNestedParallel(BaseTestCase):
         for step in steps:
             task = self.get_ready_user_tasks()[0]
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
