+++ tests/SpiffWorkflow/camunda/ResetTokenParallelMatrixTest.py
@@ -73,7 +73,7 @@ class ResetTokenTestParallelMatrix(BaseTestCase):
             if firsttaskid is None and step['taskname']=='FormB2':
                 firsttaskid = task.id
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
@@ -94,7 +94,7 @@ class ResetTokenTestParallelMatrix(BaseTestCase):
         for step in steps:
             task = self.get_ready_user_tasks()[0]
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
@@ -153,7 +153,7 @@ class ResetTokenTestParallelMatrix(BaseTestCase):
             if firsttaskid is None and step['taskname']=='FormA2':
                 firsttaskid = task.id
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
@@ -190,7 +190,7 @@ class ResetTokenTestParallelMatrix(BaseTestCase):
         for step in steps:
             task = self.get_ready_user_tasks()[0]
             self.assertEqual(step['taskname'], task.task_spec.name)
-            task.update_data({step['formvar']: step['answer']})
+            task.set_data(**{step['formvar']: step['answer']})
             self.workflow.run_task_from_id(task.id)
             self.workflow.do_engine_steps()
             if save_restore:
